Moved unused snippets to gist

+++ makeSpritesheets.ps1
@@ -13,5 +13,7 @@ foreach ($dir in $subDirs) {
 
     TexturePacker "settings.tps" --sheet "$outputDir\$folderName.webp" --data "$outputDir\$folderName.json" $fullName
 
-    Remove-Item -Recurse -Force $fullName
+    if (Test-Path "$outputDir\$folderName.webp" -and Test-Path "$outputDir\$folderName.json") {
+        Remove-Item -Recurse -Force $fullName
+    }
 }
